+++ packages/components/buttons/primary-button/src/primary-button.styles.ts
@@ -10,13 +10,11 @@ const getSizeStyles = (size: TPrimaryButtonProps['size']) => {
     case 'medium':
       return css`
         height: ${designTokens.heightForButtonAsMedium};
-        border-radius: ${designTokens.borderRadiusForButtonAsMedium};
       `;
 
     case 'big':
       return css`
         height: ${designTokens.heightForButtonAsBig};
-        border-radius: ${designTokens.borderRadiusForButtonAsBig};
       `;
 
     default:
@@ -34,7 +32,8 @@ const getButtonStyles = (
     align-items: center;
     color: ${designTokens.colorSurface};
     transition: background-color ${designTokens.transitionLinear80Ms};
-    padding: ${designTokens.paddingForButton};
+    padding: 0 ${designTokens.spacing30};
+    border-radius: ${designTokens.borderRadius4};
     ${getSizeStyles(size)}
   `;
   // "disabled" takes precendece over "active"
@@ -71,10 +70,10 @@ const getButtonStyles = (
         return [
           baseActiveStyles,
           css`
-            background-color: ${designTokens.backgroundColorForButtonAsPrimary};
+            background-color: ${designTokens.colorPrimary};
             &:focus,
             &:hover {
-              background-color: ${designTokens.backgroundColorForButtonAsPrimaryAsDefaultWhenHovered};
+              background-color: ${designTokens.colorPrimary40};
             }
           `,
         ];
@@ -134,10 +133,10 @@ const getButtonStyles = (
           background-color: ${designTokens.colorPrimary};
           &:focus,
           &:hover {
-            background-color: ${designTokens.backgroundColorForButtonAsPrimaryAsDefaultWhenHovered};
+            background-color: ${designTokens.colorPrimary40};
           }
           &:active {
-            background-color: ${designTokens.backgroundColorForButtonAsPrimaryAsDefaultWhenActive};
+            background-color: ${designTokens.colorPrimary25};
           }
         `,
       ];
@@ -148,7 +147,7 @@ const getButtonStyles = (
       return [
         baseDefaultStyles,
         css`
-          background-color: ${designTokens.backgroundColorForButtonAsPrimaryAsUrgent};
+          background-color: ${designTokens.colorWarning60};
           &:focus,
           &:hover {
             background-color: #e7680d;