+++ packages/components/inputs/creatable-select-input/src/creatable-select-input.tsx
@@ -265,6 +265,10 @@ export type TCreatableSelectInputProps = {
    */
   placeholder?: ReactSelectCreatableProps['placeholder'];
   /**
+   * Use this property to reduce the paddings of the component for a ui compact variant
+   */
+  isCondensed?: boolean;
+  /**
    * Sets the tabIndex attribute on the input
    * <br>
    * [Props from React select was used](https://react-select.com/props#creatable-props)
@@ -389,6 +393,7 @@ const CreatableSelectInput = (props: TCreatableSelectInputProps) => {
               hasError: props.hasError,
               showOptionGroupDivider: props.showOptionGroupDivider,
               menuPortalZIndex: props.menuPortalZIndex,
+              isCondensed: props.isCondensed,
               isDisabled: props.isDisabled,
               isReadOnly: props.isReadOnly,
               iconLeft: props.iconLeft,
@@ -485,6 +490,7 @@ const CreatableSelectInput = (props: TCreatableSelectInputProps) => {
           createOptionPosition={props.createOptionPosition}
           //@ts-ignore
           iconLeft={props.iconLeft}
+          isCondensed={props.isCondensed}
         />
       </div>
     </Constraints.Horizontal>