+++ packages/components/inputs/date-time-input/src/date-time-input.story.js
@@ -50,6 +50,7 @@ class DateTimeInputStory extends Component {
           id={text('id', '')}
           name={text('name', '')}
           placeholder={placeholder === '' ? undefined : placeholder}
+          isCondensed={boolean('isCondensed', false)}
           isDisabled={boolean('isDisabled', false)}
           isReadOnly={boolean('isReadOnly', false)}
           hasError={boolean('hasError', false)}