+++ packages/components/inputs/localized-money-input/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @commercetools-uikit/localized-money-input
 
+## 19.2.0
+
+### Minor Changes
+
+- [#2785](https://github.com/commercetools/ui-kit/pull/2785) [`3a47d4a`](https://github.com/commercetools/ui-kit/commit/3a47d4a673b4642c7a697eef85afbc41ca6cc526) Thanks [@ddouglasz](https://github.com/ddouglasz)! - Remove unused old theme and token reference
+
+- [#2798](https://github.com/commercetools/ui-kit/pull/2798) [`0798b97`](https://github.com/commercetools/ui-kit/commit/0798b97e9bb6f8e843a48420f2e1b6e9a53a9b03) Thanks [@Sarah4VT](https://github.com/Sarah4VT)! - Add support for isCondensed property on LocalizedMoneyInput component
+
+### Patch Changes
+
+- Updated dependencies [[`d72e043`](https://github.com/commercetools/ui-kit/commit/d72e043db3f31c330c4a8c34a7b3fbd623387fb3), [`3a47d4a`](https://github.com/commercetools/ui-kit/commit/3a47d4a673b4642c7a697eef85afbc41ca6cc526), [`0798b97`](https://github.com/commercetools/ui-kit/commit/0798b97e9bb6f8e843a48420f2e1b6e9a53a9b03)]:
+  - @commercetools-uikit/input-utils@19.2.0
+  - @commercetools-uikit/spacings-stack@19.2.0
+  - @commercetools-uikit/flat-button@19.2.0
+  - @commercetools-uikit/select-utils@19.2.0
+  - @commercetools-uikit/money-input@19.2.0
+  - @commercetools-uikit/constraints@19.2.0
+  - @commercetools-uikit/messages@19.2.0
+  - @commercetools-uikit/tooltip@19.2.0
+  - @commercetools-uikit/icons@19.2.0
+  - @commercetools-uikit/localized-utils@19.2.0
+  - @commercetools-uikit/hooks@19.2.0
+  - @commercetools-uikit/utils@19.2.0
+  - @commercetools-uikit/design-system@19.2.0
+
 ## 19.1.0
 
 ### Patch Changes